+++ src/main/java/Lotto.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Lotto {
     private ArrayList<Integer> numbers = new ArrayList<>();
     private ArrayList<Integer> lotto = new ArrayList<>();
-    private ArrayList<Integer> counts = new ArrayList<>();
+    static ArrayList<Integer> counts = new ArrayList<>();
 
     public Lotto() {
         for(int i = 0; i < 46; i++) {
@@ -23,7 +23,7 @@ public class Lotto {
         }
     }
 
-    public Integer checkNumber(List<String> rightNum) {
+    public ArrayList<Integer> checkNumber(List<String> rightNum) {
         int count = 0;
         for(int i = 0; i < rightNum.size(); i++) {
             if(lotto.contains(Integer.parseInt(rightNum.get(i))))
@@ -31,31 +31,10 @@ public class Lotto {
         }
         counts.add(count);
         System.out.print("count is " + count);
-        return count;
+        return counts;
     }
 
-    public void money(ArrayList<Integer> counts) {
-        for(int i = 0; i < counts.size(); i++) {
-            int num = 0;
-            if(counts.get(i) == 3) {
-                num++;
-                System.out.println("3개 일치 (5000원)- " + num);
-            }
 
-            if(counts.get(i) == 4) {
-                num++;
-                System.out.println("4개 일치 (50000원)- " + num);
-            }
 
-            if(counts.get(i) == 5) {
-                num++;
-                System.out.println("5개 일치 (1500000원)- " + num);
-            }
 
-            if(counts.get(i) == 6) {
-                num++;
-                System.out.println("6개 일치 (2000000000원)- " + num);
-            }
-        }
-    }
 }
